requirements.txt
httpx>=0.28
requests>=2.32
python-telegram-bot==21.11.1
apscheduler>=3.11
pydantic>=2.12
fastapi>=0.135
uvicorn>=0.42
python-dotenv>=1.0
